+++ stepik-union/src/main/resources/META-INF/pycharm.xml
@@ -3,6 +3,6 @@
         <treeStructureProvider implementation="org.stepik.plugin.projectView.PyCharmTreeStructureProvider"/>
         <navbar implementation="org.stepik.plugin.projectView.PyCharmNavBarModelExtension"/>
         <moduleType id="JAVA_MODULE" implementationClass="org.stepik.plugin.projectWizard.compatibility.PyCharmJavaModuleType"/>
-        <directoryProjectGenerator implementation="org.stepik.plugin.projectWizard.pycharm.StepikPyProjectGenerator"/>
+        <directoryProjectGenerator implementation="org.stepik.plugin.projectWizard.StepikPyProjectGenerator"/>
     </extensions>
 </idea-plugin>
